Move RuntimeSettings into utils/config/settings.hpp
